Added comments on the article detail page and moved some functions to the services.py
