+++ demo/views.py
@@ -3,7 +3,7 @@ from django.http import HttpResponse
 from django.shortcuts import render, get_object_or_404
 from datetime import datetime
 from django.utils import timezone
-from demo.models import Article
+
 
 
 def homepage_view(request):
@@ -46,14 +46,6 @@ def topic_unsubscribe_view(request, topic: str):
     return HttpResponse(f'This page is to confirm unsubscription from the topic {topic}')
 
 
-def article_list(request):
-    articles = Article.objects.filter(created__lte=timezone.now()).order_by('created')
-    return render(request, 'main_page.html', {"articles": articles})
-
-
-def article_detail(request, pk):
-    article = get_object_or_404(Article, pk=pk)
-    return render(request, "article_detail.html", {"article": article})
